Moved logic for logout and refreshing token to redux
