+++ src/store/reducer.js
@@ -79,20 +79,27 @@ else {
 
 // Initial State
 const initialState = {
+
   Theme: { ...presetTheme },
   themeDict: themeDictionary,
+
   signUpMessage: '',
   signUpFollowLink: '',
+
   signInMessage: '',
   signInFollowLink: '',
+
   forgotPassMessage: '',
   forgotPassFollowLink: '',
+
   resendEmailMessage: '',
   resendEmailFollowLink: ''
+
 }
 
 function reducer(state = initialState, action) {
   switch (action.type) {
+
     case actionTypes.SET_THEME:
       localStorage.setItem("themeSet", JSON.stringify(action.payload.theme))
       if (action.payload.theme === "Auto-change")
@@ -105,30 +112,35 @@ function reducer(state = initialState, action) {
           ...state,
           Theme: { ...themeDictionary[action.payload.theme] }
         }
+
     case actionTypes.SET_SIGNUP_MESSAGE:
       return {
         ...state,
         signUpMessage: action.payload.message,
         signUpFollowLink: action.payload.link
       }
+
     case actionTypes.SET_SIGNIN_MESSAGE:
       return {
         ...state,
         signInMessage: action.payload.message,
         signInFollowLink: action.payload.link
       }
+
     case actionTypes.SET_FORGOT_PASS_MESSAGE:
       return {
         ...state,
         forgotPassMessage: action.payload.message,
         forgotPassFollowLink: action.payload.link
       }
+
     case actionTypes.SET_RESEND_EMAIL_MESSAGE:
       return {
         ...state,
         resendEmailMessage: action.payload.message,
         resendEmailFollowLink: action.payload.link
       }
+
     default:
       return {
         ...state
