+++ src/views/Movie.js
@@ -1,12 +1,12 @@
 import * as React from "react";
 import * as PropTypes from "prop-types";
 import axios from 'axios';
-import createAuthRefreshInterceptor from 'axios-auth-refresh';
-
+import { connect } from "react-redux";
 import AppBarButton from "react-uwp/AppBarButton";
 import Button from "react-uwp/Button";
+import * as actionCreators from '../store/actions/actionCreators';
 
-export default class Movie extends React.Component {
+class Movie extends React.Component {
 
   constructor(props) {
     super(props);
@@ -34,27 +34,6 @@ export default class Movie extends React.Component {
     this.getData = this.getData.bind(this);
     this.handleAddSeen = this.handleAddSeen.bind(this);
     this.handleAddBucket = this.handleAddBucket.bind(this);
-    this.refreshToken = this.refreshToken.bind(this);
-  }
-
-  refreshToken() {
-    const refreshOptions = {
-      'method': 'POST',
-      'url': `${process.env.REACT_APP_DB_HOST}/auth/refreshToken`,
-      'headers': {
-        'Authorization': localStorage.getItem("refresh_token")
-      }
-    }
-
-    const refreshAuthLogic = failedRequest => axios(refreshOptions)
-      .then(tokenRefreshResponse => {
-        localStorage.setItem('access_token', tokenRefreshResponse.data.access_token);
-        localStorage.setItem('refresh_token', tokenRefreshResponse.data.refresh_token);
-        failedRequest.response.config.headers['Authorization'] = tokenRefreshResponse.data.access_token;
-        return Promise.resolve();
-      });
-
-    return refreshAuthLogic;
   }
 
   getData() {
@@ -123,9 +102,7 @@ export default class Movie extends React.Component {
         }
       }
 
-      const refreshAuthLogic = this.refreshToken();
-
-      createAuthRefreshInterceptor(axios, refreshAuthLogic);
+      this.props.registerRefreshToken(axios);
 
       axios(axiosOptions)
         .then(response => {
@@ -156,9 +133,7 @@ export default class Movie extends React.Component {
         }
       }
 
-      const refreshAuthLogic = this.refreshToken();
-
-      createAuthRefreshInterceptor(axios, refreshAuthLogic);
+      this.props.registerRefreshToken(axios);
 
       axios(axiosOptions)
         .then(response => {
@@ -269,3 +244,17 @@ export default class Movie extends React.Component {
     );
   }
 }
+
+const mapStateToProps = null/*state => {
+  return {
+
+  }
+}*/
+
+const mapDispatchToProps = dispatch => {
+  return {
+    registerRefreshToken: (axiosInstance) => dispatch(actionCreators.registerRefreshToken(axiosInstance))
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Movie);
